Code Updated for Procedures
I am pretty sure that the code is working with procedures.

Please note, the included vm.exe is for windows and does not do the bars
properly. The version that me and Jacob worked together on did them
correctly, but I'd have to rewrite our entire vm.c to fix this.